HollowPurple
Sorry, Amanai. I'm not even angry over you right now. I bear no grudge against anyone. It's just that the world feels so, so wonderful right now.

Throughout Heaven and Earth, I alone am the honored one.

The merit of having a technique that's passed down for generations is having a user's manual.

The demerit is that information about the technique is easily leaked.

You were a member of the Zen'in clan, weren't you? That's why you know so much about the Limitless technique.

However, even in the Gojo clan, only a scant few know about this.

Take the amplified and the reversal, then smash together those two different expressions of infinity to create and push out imaginary mass.

Imaginary Technique: Purple.